Make it faster to develop webpack configuration
WPS will handle retarting the sever when you change a bundled file. It's
oblivious to changes made to webpack configuration, though, and you have
to restart the WPS whenever you change something. The process was
automated by addind nodemon and making it watch for changes in webpack
configuration files - yes, configuration fileS for we're going to
compose the configuration.